requirements.txt
python-telegram-bot==12.1.0
requests==2.22.0
urllib3==1.25.3
sqlalchemy==1.3.8
psycopg2==2.8.3
sqlalchemy-migrate==0.12.0
